gofmt code formatter and refactoring imports
