+++ backend/src/logs/logs.log
@@ -460,3 +460,6 @@
 {"level":"info","ts":"2023-10-03T10:39:53.706+0330","msg":"","AppName":"MyApp","LoggerName":"ZapLogger","Latency":0.000577,"Method":"POST","BodySize":79,"Path":"/api/v1/power/calc/","RequestBody":"\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\r\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\r\n\n\n\n\n\r\n\r\n\n\n\n\n\n\n\r\n\r\n\n\n\n\n\r\n{\r\n  \"cpuId\": 1,\r\n  \"graphicCount\": 20,\r\n  \"graphicPower\": 13,\r\n\n  \"hardDrivePower\": 12,\r\n  \"motherboardPower\": 3,\r\n  \"opticalDriveCount\": 2,\r\n  \"opticalDrivePower\": 12,\r\n  \"ramCount\": 2,\r\n  \"ramPower\": 18,\r\n  \"ssdCount\": 1,\r\n  \"ssdPower\": 3\r\n}\r\n\r\n","Category":"RequestResponse","SubCategory":"Api","ClientIp":"127.0.0.1"}
 {"level":"info","ts":"2023-10-03T10:46:21.385+0330","msg":"postgres started","AppName":"MyApp","LoggerName":"ZapLogger","Category":"Postgres","SubCategory":"Startup"}
 {"level":"info","ts":"2023-10-03T10:46:22.881+0330","msg":"","AppName":"MyApp","LoggerName":"ZapLogger","Method":"POST","BodySize":328,"Path":"/api/v1/ram/all/","RequestBody":"\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\r\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\r\n{\r\n    \"page_number\": 0,\r\n  \"page_size\": 0\r\n}\r\n\r\n\r\n\n\n\n\n\n\n\r\n\r\n\n\n\n\n\r\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n\r\n","Category":"RequestResponse","SubCategory":"Api","ClientIp":"127.0.0.1","Latency":0.1064096}
+{"level":"fatal","ts":"2023-10-03T12:09:47.410+0330","msg":"cant init db","AppName":"MyApp","LoggerName":"ZapLogger","Category":"Internal","SubCategory":"Startup","stacktrace":"main.main\n\tD:/Code/PSU-Calculator/backend/src/cmd/main.go:18\nruntime.main\n\tC:/Program Files/Go/src/runtime/proc.go:250"}
+{"level":"fatal","ts":"2023-10-03T12:10:10.719+0330","msg":"cant init db","AppName":"MyApp","LoggerName":"ZapLogger","Category":"Internal","SubCategory":"Startup","stacktrace":"main.main\n\tD:/Code/PSU-Calculator/backend/src/cmd/main.go:18\nruntime.main\n\tC:/Program Files/Go/src/runtime/proc.go:250"}
+{"level":"info","ts":"2023-10-03T12:11:39.332+0330","msg":"postgres started","AppName":"MyApp","LoggerName":"ZapLogger","Category":"Postgres","SubCategory":"Startup"}
