+++ backend/src/services/base_service.go
@@ -9,13 +9,14 @@ import (
 	"strings"
 	"time"
 
-	"github.com/Vigiatonet/PSU-Calculator/src/api/dto"
-	"github.com/Vigiatonet/PSU-Calculator/src/common"
-	"github.com/Vigiatonet/PSU-Calculator/src/config"
-	"github.com/Vigiatonet/PSU-Calculator/src/constants"
-	"github.com/Vigiatonet/PSU-Calculator/src/data/db"
-	"github.com/Vigiatonet/PSU-Calculator/src/data/models"
-	"github.com/Vigiatonet/PSU-Calculator/src/pkg/logging"
+	"github.com/Vigiatonet/PSU-Calculator/api/dto"
+	"github.com/Vigiatonet/PSU-Calculator/common"
+	"github.com/Vigiatonet/PSU-Calculator/config"
+	"github.com/Vigiatonet/PSU-Calculator/constants"
+	"github.com/Vigiatonet/PSU-Calculator/data/db"
+	"github.com/Vigiatonet/PSU-Calculator/data/models"
+	"github.com/Vigiatonet/PSU-Calculator/pkg/logging"
+
 	"gorm.io/gorm"
 )
