+++ backend/src/services/token_service.go
@@ -3,14 +3,14 @@ package services
 import (
 	"time"
 
-	"github.com/Vigiatonet/PSU-Calculator/src/api/dto"
-	"github.com/Vigiatonet/PSU-Calculator/src/config"
-	"github.com/Vigiatonet/PSU-Calculator/src/constants"
-	"github.com/Vigiatonet/PSU-Calculator/src/data/cache"
-	"github.com/Vigiatonet/PSU-Calculator/src/data/db"
-	"github.com/Vigiatonet/PSU-Calculator/src/data/models"
-	"github.com/Vigiatonet/PSU-Calculator/src/pkg/logging"
-	"github.com/Vigiatonet/PSU-Calculator/src/pkg/service_errors"
+	"github.com/Vigiatonet/PSU-Calculator/api/dto"
+	"github.com/Vigiatonet/PSU-Calculator/config"
+	"github.com/Vigiatonet/PSU-Calculator/constants"
+	"github.com/Vigiatonet/PSU-Calculator/data/cache"
+	"github.com/Vigiatonet/PSU-Calculator/data/db"
+	"github.com/Vigiatonet/PSU-Calculator/data/models"
+	"github.com/Vigiatonet/PSU-Calculator/pkg/logging"
+	"github.com/Vigiatonet/PSU-Calculator/pkg/service_errors"
 	"github.com/golang-jwt/jwt/v4"
 )
