+++ backend/src/services/user_service.go
@@ -4,13 +4,13 @@ import (
 	"context"
 	"database/sql"
 
-	"github.com/Vigiatonet/PSU-Calculator/src/api/dto"
-	"github.com/Vigiatonet/PSU-Calculator/src/config"
-	"github.com/Vigiatonet/PSU-Calculator/src/constants"
-	"github.com/Vigiatonet/PSU-Calculator/src/data/db"
-	"github.com/Vigiatonet/PSU-Calculator/src/data/models"
-	"github.com/Vigiatonet/PSU-Calculator/src/pkg/logging"
-	"github.com/Vigiatonet/PSU-Calculator/src/pkg/service_errors"
+	"github.com/Vigiatonet/PSU-Calculator/api/dto"
+	"github.com/Vigiatonet/PSU-Calculator/config"
+	"github.com/Vigiatonet/PSU-Calculator/constants"
+	"github.com/Vigiatonet/PSU-Calculator/data/db"
+	"github.com/Vigiatonet/PSU-Calculator/data/models"
+	"github.com/Vigiatonet/PSU-Calculator/pkg/logging"
+	"github.com/Vigiatonet/PSU-Calculator/pkg/service_errors"
 	"golang.org/x/crypto/bcrypt"
 	"gorm.io/gorm"
 )
